% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guttmacher_theme.R
\name{get_categorical_palette}
\alias{get_categorical_palette}
\title{Returns categorical palette}
\usage{
get_categorical_palette(num_categories = 4)
}
\arguments{
\item{num_categories}{An integer representing the number of categorical variables.}
}
\value{
categorical_palette A vector of categorical colors.
}
\description{
Returns categorical palette
}
\examples{
get_categorical_palette()
get_categorical_palette(num_categories=2)
}
