% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guttmacher_theme.R
\name{get_divergent_palette}
\alias{get_divergent_palette}
\title{function that returns divergent palette}
\usage{
get_divergent_palette()
}
\value{
divergent_palette
}
\description{
function that returns divergent palette
}
\examples{
get_divergent_palette()
}
