% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guttmacher_theme.R
\name{guttmacher_theme}
\alias{guttmacher_theme}
\title{Guttmacher ggplot theme}
\usage{
guttmacher_theme(base_family = "Arial", base_size = 11, show_legend = TRUE)
}
\arguments{
\item{show_legend}{}
}
\description{
Guttmacher ggplot theme
}
\examples{
guttmacher_theme()
guttmacher_theme(base_size = 15)
}
