% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guttmacher_theme.R
\name{get_categorical_palette}
\alias{get_categorical_palette}
\title{function that returns categorical palette}
\usage{
get_categorical_palette(num_categories = 4)
}
\arguments{
\item{num_categories}{}
}
\value{
categorical_palette
}
\description{
function that returns categorical palette
}
\examples{
get_categorical_palette()
get_categorical_palette(num_categories=2)
}
