% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guttmacher_theme.R
\name{get_divergent_categorical_palette}
\alias{get_divergent_categorical_palette}
\title{function that returns divergent categorical palette}
\usage{
get_divergent_categorical_palette()
}
\value{
divergent_categorical_palette
}
\description{
function that returns divergent categorical palette
}
\examples{
get_divergent_categorical_palette()
}
