% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guttmacher_theme.R
\name{get_sequential_categorical_palette}
\alias{get_sequential_categorical_palette}
\title{function that returns sequential categorical palette}
\usage{
get_sequential_categorical_palette(base_color = "orange", num_categories = 6)
}
\arguments{
\item{num_categories}{}
}
\value{
sequential_categorical_palette
}
\description{
function that returns sequential categorical palette
}
\examples{
get_sequential_categorical_palette()
get_sequential_categorical_palette(base_color="blue")
}
