% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guttmacher_theme.R
\name{get_sequential_palette}
\alias{get_sequential_palette}
\title{function that returns sequential palette}
\usage{
get_sequential_palette(base_color = "orange")
}
\arguments{
\item{base_color}{}
}
\value{
sequential_palette
}
\description{
function that returns sequential palette
}
\examples{
get_sequential_palette()
get_sequential_palette(base_color="violet")
}
