% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guttmacher_theme.R
\name{guttmacher_theme}
\alias{guttmacher_theme}
\title{Guttmacher ggplot theme}
\usage{
guttmacher_theme(
  base_family = "Arial",
  base_size = 11,
  show_legend = TRUE,
  legend_border = TRUE
)
}
\arguments{
\item{base_family}{A string value representing base family name.}

\item{base_size}{A numeric value representing base font size.}

\item{show_legend}{A Boolean value indicating whether to show the legend or not.}
}
\value{
A ggplot theme that fits Guttmacher's branding guidelines.
}
\description{
Guttmacher ggplot theme
}
\examples{
guttmacher_theme()
guttmacher_theme(base_size = 15)
}
